feat: Basic routing
Have completed routing between the index page (where all of the books are listed) and the search page. User can starts on the page where all of there books are listed and has the option to add more books. The add more books link will take them to the /seach route. The user can either push the browser back button to get back to the main page or they can push the close search button to get there (which I also added a react router link to).